Move the weather icons down to infrastructure
